+++ packages/@expo/cli/src/export/fork-bundleAsync.ts
@@ -31,6 +31,7 @@ export type BundleOptions = {
   platform: 'android' | 'ios' | 'web';
   dev?: boolean;
   minify?: boolean;
+  bytecode: boolean;
   sourceMapUrl?: string;
   sourcemaps?: boolean;
 };
@@ -70,6 +71,7 @@ export async function createBundlesAsync(
     platforms: Platform[];
     dev?: boolean;
     minify?: boolean;
+    bytecode: boolean;
     sourcemaps?: boolean;
     entryPoint?: string;
   }
@@ -93,6 +95,7 @@ export async function createBundlesAsync(
         bundleOptions.entryPoint ?? getEntryWithServerRoot(projectRoot, { platform, pkg }),
       sourcemaps: bundleOptions.sourcemaps,
       minify: bundleOptions.minify,
+      bytecode: bundleOptions.bytecode,
       dev: bundleOptions.dev,
     }))
   );
@@ -151,15 +154,16 @@ async function bundleProductionMetroClientAsync(
       ...Server.DEFAULT_BUNDLE_OPTIONS,
       sourceMapUrl: bundle.sourceMapUrl,
       ...getMetroDirectBundleOptionsForExpoConfig(projectRoot, expoConfig, {
+        minify: bundle.minify,
         mainModuleName: bundle.entryPoint,
         platform: bundle.platform,
         mode: bundle.dev ? 'development' : 'production',
         engine: isHermes ? 'hermes' : undefined,
         serializerIncludeMaps: bundle.sourcemaps,
+        bytecode: bundle.bytecode && isHermes,
         // Bundle splitting on web-only for now.
         // serializerOutput: bundle.platform === 'web' ? 'static' : undefined,
         serializerOutput: 'static',
-        serializerIncludeBytecode: isHermes,
         isExporting: true,
       }),
       bundleType: 'bundle',